Create Cabal-tree-diff package
Move Cabal instances from cabal-install TreeDiffInstances into it.
Remove defaultToExprViaShow to get better diffs and dumps

Also rename Cabal-QuickCheck (proper case), and add GenericArbitrary
module there.
