+++ Cabal/Cabal-QuickCheck/src/Test/QuickCheck/Instances/Cabal.hs
@@ -1,6 +1,4 @@
 {-# LANGUAGE CPP              #-}
-{-# LANGUAGE FlexibleContexts #-}
-{-# LANGUAGE GADTs            #-}
 {-# LANGUAGE TypeOperators #-}
 {-# OPTIONS_GHC -fno-warn-orphans #-}
 module Test.QuickCheck.Instances.Cabal () where
@@ -10,8 +8,6 @@ import Data.Char                  (isAlphaNum, isDigit)
 import Data.List                  (intercalate)
 import Distribution.Utils.Generic (lowercase)
 import Test.QuickCheck
-
-import GHC.Generics
 
 import Distribution.CabalSpecVersion
 import Distribution.Compiler
@@ -31,6 +27,8 @@ import Distribution.Types.UnqualComponentName
 import Distribution.Types.VersionRange.Internal
 import Distribution.Verbosity
 import Distribution.Version
+
+import Test.QuickCheck.GenericArbitrary
 
 #if !MIN_VERSION_base(4,8,0)
 import Control.Applicative (pure, (<$>), (<*>))
@@ -342,38 +340,3 @@ shortListOf1 :: Int -> Gen a -> Gen [a]
 shortListOf1 bound gen = sized $ \n -> do
     k <- choose (1, 1 `max` ((n `div` 2) `min` bound))
     vectorOf k gen
-
--------------------------------------------------------------------------------
--- Generic Arbitrary
--------------------------------------------------------------------------------
-
--- Generic arbitary for non-recursive types
-genericArbitrary :: (Generic a, GArbitrary (Rep a)) => Gen a
-genericArbitrary = fmap to garbitrary
-
-class GArbitrary f where
-    garbitrary :: Gen (f ())
-
-class GArbitrarySum f where
-    garbitrarySum :: [Gen (f ())]
-
-class GArbitraryProd f where
-    garbitraryProd :: Gen (f ())
-
-instance (GArbitrarySum f, i ~ D) => GArbitrary (M1 i c f) where
-    garbitrary = fmap M1 (oneof garbitrarySum)
-
-instance (GArbitraryProd f, i ~ C) => GArbitrarySum (M1 i c f) where
-    garbitrarySum = [fmap M1 garbitraryProd]
-
-instance (GArbitrarySum f, GArbitrarySum g) => GArbitrarySum (f :+: g) where
-    garbitrarySum = map (fmap L1) garbitrarySum ++ map (fmap R1) garbitrarySum
-
-instance (GArbitraryProd f, i ~ S) => GArbitraryProd (M1 i c f) where
-    garbitraryProd = fmap M1 garbitraryProd
-
-instance (GArbitraryProd f, GArbitraryProd g) => GArbitraryProd (f :*: g) where
-    garbitraryProd = liftA2 (:*:) garbitraryProd garbitraryProd
-
-instance (Arbitrary a) => GArbitraryProd (K1 i a) where
-    garbitraryProd = fmap K1 arbitrary
