+++ Cabal/Cabal-tree-diff/Cabal-tree-diff.cabal
@@ -1,0 +1,21 @@
+cabal-version: 2.2
+name:          Cabal-tree-diff
+version:       3.3.0.0
+synopsis:      QuickCheck instances for types in Cabal
+category:      Testing
+description:   Provides tree-diff ToExpr instances for some types in Cabal
+
+library
+  default-language: Haskell2010
+  hs-source-dirs:   src
+  ghc-options:      -Wall
+  build-depends:
+    , base
+    , Cabal      ^>=3.3.0.0
+    , tree-diff  ^>=0.1
+
+  exposed-modules:  Data.TreeDiff.Instances.Cabal
+  other-modules:
+    Data.TreeDiff.Instances.CabalLanguage
+    Data.TreeDiff.Instances.CabalSPDX
+    Data.TreeDiff.Instances.CabalVersion
