+++ Cabal/tests/ParserTests/ipi/Includes2.expr
@@ -1,7 +1,11 @@
 InstalledPackageInfo
-  {abiDepends = [`AbiDependency {depUnitId = UnitId "base-4.10.1.0", depAbiHash = AbiHash "35a7f6be752ee4f7385cb5bf28677879"}`,
-                 `AbiDependency {depUnitId = UnitId "Includes2-0.1.0.0-inplace-mysql", depAbiHash = AbiHash "inplace"}`],
-   abiHash = `AbiHash "inplace"`,
+  {abiDepends = [AbiDependency
+                   {depAbiHash = AbiHash "35a7f6be752ee4f7385cb5bf28677879",
+                    depUnitId = UnitId "base-4.10.1.0"},
+                 AbiDependency
+                   {depAbiHash = AbiHash "inplace",
+                    depUnitId = UnitId "Includes2-0.1.0.0-inplace-mysql"}],
+   abiHash = AbiHash "inplace",
    author = "Edward Z. Yang",
    category = "",
    ccOptions = [],
@@ -9,11 +13,12 @@ InstalledPackageInfo
    copyright = "",
    cxxOptions = [],
    dataDir = "/home/travis/build/haskell/cabal/cabal-testsuite/PackageTests/Backpack/Includes2",
-   depends = [`UnitId "base-4.10.1.0"`,
-              `UnitId "Includes2-0.1.0.0-inplace-mysql"`],
+   depends = [UnitId "base-4.10.1.0",
+              UnitId "Includes2-0.1.0.0-inplace-mysql"],
    description = "",
    exposed = False,
-   exposedModules = [`ExposedModule {exposedName = ModuleName "Mine", exposedReexport = Nothing}`],
+   exposedModules = [ExposedModule
+                       {exposedName = ModuleName "Mine", exposedReexport = Nothing}],
    extraGHCiLibraries = [],
    extraLibraries = [],
    frameworkDirs = [],
@@ -27,14 +32,15 @@ InstalledPackageInfo
    includeDirs = [],
    includes = [],
    indefinite = False,
-   installedComponentId_ = `ComponentId ""`,
-   installedUnitId = `UnitId "Includes2-0.1.0.0-inplace-mylib+3gY9SyjX86dBypHcOaev1n"`,
+   installedComponentId_ = ComponentId "",
+   installedUnitId = UnitId
+                       "Includes2-0.1.0.0-inplace-mylib+3gY9SyjX86dBypHcOaev1n",
    instantiatedWith = [_×_
-                         `ModuleName "Database"`
+                         (ModuleName "Database")
                          (OpenModule
                             (DefiniteUnitId
-                               (DefUnitId `UnitId "Includes2-0.1.0.0-inplace-mysql"`))
-                            `ModuleName "Database.MySQL"`)],
+                               (DefUnitId (UnitId "Includes2-0.1.0.0-inplace-mysql")))
+                            (ModuleName "Database.MySQL"))],
    ldOptions = [],
    libVisibility = LibraryVisibilityPrivate,
    libraryDirs = ["/home/travis/build/haskell/cabal/cabal-testsuite/PackageTests/Backpack/Includes2/cabal-internal.dist/work/./dist/build/x86_64-linux/ghc-8.2.2/Includes2-0.1.0.0/l/mylib/Includes2-0.1.0.0-inplace-mylib+3gY9SyjX86dBypHcOaev1n/build/Includes2-0.1.0.0-inplace-mylib+3gY9SyjX86dBypHcOaev1n"],
@@ -43,10 +49,10 @@ InstalledPackageInfo
    maintainer = "ezyang@cs.stanford.edu",
    pkgRoot = Nothing,
    pkgUrl = "",
-   sourceLibName = LSubLibName `UnqualComponentName "mylib"`,
+   sourceLibName = LSubLibName (UnqualComponentName "mylib"),
    sourcePackageId = PackageIdentifier
-                       {pkgName = `PackageName "Includes2"`,
-                        pkgVersion = `mkVersion [0,1,0,0]`},
+                       {pkgName = PackageName "Includes2",
+                        pkgVersion = mkVersion [0, 1, 0, 0]},
    stability = "",
    synopsis = "",
    trusted = False}
