+++ Cabal/tests/ParserTests/ipi/internal-preprocessor-test.expr
@@ -1,6 +1,6 @@
 InstalledPackageInfo
   {abiDepends = [],
-   abiHash = `AbiHash ""`,
+   abiHash = AbiHash "",
    author = "Mikhail Glushenkov",
    category = "Testing",
    ccOptions = [],
@@ -8,10 +8,11 @@ InstalledPackageInfo
    copyright = "",
    cxxOptions = [],
    dataDir = "/home/ogre/Documents/other-haskell/cabal/cabal-testsuite/PackageTests/CustomPreProcess",
-   depends = [`UnitId "base-4.8.2.0-0d6d1084fbc041e1cded9228e80e264d"`],
+   depends = [UnitId "base-4.8.2.0-0d6d1084fbc041e1cded9228e80e264d"],
    description = "See https://github.com/haskell/cabal/issues/1541#issuecomment-30155513",
    exposed = True,
-   exposedModules = [`ExposedModule {exposedName = ModuleName "A", exposedReexport = Nothing}`],
+   exposedModules = [ExposedModule
+                       {exposedName = ModuleName "A", exposedReexport = Nothing}],
    extraGHCiLibraries = [],
    extraLibraries = [],
    frameworkDirs = [],
@@ -25,23 +26,23 @@ InstalledPackageInfo
    includeDirs = [],
    includes = [],
    indefinite = False,
-   installedComponentId_ = `ComponentId ""`,
-   installedUnitId = `UnitId "internal-preprocessor-test-0.1.0.0"`,
+   installedComponentId_ = ComponentId "",
+   installedUnitId = UnitId "internal-preprocessor-test-0.1.0.0",
    instantiatedWith = [],
    ldOptions = [],
    libVisibility = LibraryVisibilityPublic,
    libraryDirs = ["/home/ogre/Documents/other-haskell/cabal/cabal-testsuite/PackageTests/CustomPreProcess/setup.dist/work/dist/build",
                   "/home/ogre/Documents/other-haskell/cabal/cabal-testsuite/PackageTests/CustomPreProcess/setup.dist/work/dist/build"],
    libraryDynDirs = [],
-   license = Right (GPL (Just `mkVersion [3]`)),
+   license = Right (GPL (Just (mkVersion [3]))),
    maintainer = "mikhail.glushenkov@gmail.com",
    pkgRoot = Just
                "/home/ogre/Documents/other-haskell/cabal/cabal-testsuite/PackageTests/CustomPreProcess/setup.dist/work/dist",
    pkgUrl = "",
    sourceLibName = LMainLibName,
    sourcePackageId = PackageIdentifier
-                       {pkgName = `PackageName "internal-preprocessor-test"`,
-                        pkgVersion = `mkVersion [0,1,0,0]`},
+                       {pkgName = PackageName "internal-preprocessor-test",
+                        pkgVersion = mkVersion [0, 1, 0, 0]},
    stability = "",
    synopsis = "Internal custom preprocessor example.",
    trusted = False}
