+++ Cabal/tests/ParserTests/regressions/mixin-3.expr
@@ -1,20 +1,20 @@
 GenericPackageDescription
   {condBenchmarks = [],
    condExecutables = [_×_
-                        `UnqualComponentName "str-example"`
+                        (UnqualComponentName "str-example")
                         CondNode
                           {condTreeComponents = [],
                            condTreeConstraints = [Dependency
-                                                    `PackageName "base"`
-                                                    (OrLaterVersion `mkVersion [0]`)
+                                                    (PackageName "base")
+                                                    (OrLaterVersion (mkVersion [0]))
                                                     (Set.fromList [LMainLibName]),
                                                   Dependency
-                                                    `PackageName "str-string"`
-                                                    (OrLaterVersion `mkVersion [0]`)
+                                                    (PackageName "str-string")
+                                                    (OrLaterVersion (mkVersion [0]))
                                                     (Set.fromList [LMainLibName]),
                                                   Dependency
-                                                    `PackageName "str-bytestring"`
-                                                    (OrLaterVersion `mkVersion [0]`)
+                                                    (PackageName "str-bytestring")
+                                                    (OrLaterVersion (mkVersion [0]))
                                                     (Set.fromList [LMainLibName])],
                            condTreeData = Executable
                                             {buildInfo = BuildInfo
@@ -49,7 +49,14 @@ GenericPackageDescription
                                                             installIncludes = [],
                                                             jsSources = [],
                                                             ldOptions = [],
-                                                            mixins = [`Mixin {mixinPackageName = PackageName "str", mixinIncludeRenaming = IncludeRenaming {includeProvidesRn = HidingRenaming [ModuleName "Foo"], includeRequiresRn = DefaultRenaming}}`],
+                                                            mixins = [Mixin
+                                                                        {mixinIncludeRenaming = IncludeRenaming
+                                                                                                  {includeProvidesRn = HidingRenaming
+                                                                                                                         [ModuleName
+                                                                                                                            "Foo"],
+                                                                                                   includeRequiresRn = DefaultRenaming},
+                                                                         mixinPackageName = PackageName
+                                                                                              "str"}],
                                                             oldExtensions = [],
                                                             options = PerCompilerFlavor [] [],
                                                             otherExtensions = [],
@@ -60,25 +67,31 @@ GenericPackageDescription
                                                             sharedOptions = PerCompilerFlavor [] [],
                                                             staticOptions = PerCompilerFlavor [] [],
                                                             targetBuildDepends = [Dependency
-                                                                                    `PackageName "base"`
+                                                                                    (PackageName
+                                                                                       "base")
                                                                                     (OrLaterVersion
-                                                                                       `mkVersion [0]`)
+                                                                                       (mkVersion
+                                                                                          [0]))
                                                                                     (Set.fromList
                                                                                        [LMainLibName]),
                                                                                   Dependency
-                                                                                    `PackageName "str-string"`
+                                                                                    (PackageName
+                                                                                       "str-string")
                                                                                     (OrLaterVersion
-                                                                                       `mkVersion [0]`)
+                                                                                       (mkVersion
+                                                                                          [0]))
                                                                                     (Set.fromList
                                                                                        [LMainLibName]),
                                                                                   Dependency
-                                                                                    `PackageName "str-bytestring"`
+                                                                                    (PackageName
+                                                                                       "str-bytestring")
                                                                                     (OrLaterVersion
-                                                                                       `mkVersion [0]`)
+                                                                                       (mkVersion
+                                                                                          [0]))
                                                                                     (Set.fromList
                                                                                        [LMainLibName])],
                                                             virtualModules = []},
-                                             exeName = `UnqualComponentName "str-example"`,
+                                             exeName = UnqualComponentName "str-example",
                                              exeScope = ExecutablePublic,
                                              modulePath = "Main.hs"}}],
    condForeignLibs = [],
@@ -109,8 +122,7 @@ GenericPackageDescription
                            licenseRaw = Left NONE,
                            maintainer = "",
                            package = PackageIdentifier
-                                       {pkgName = `PackageName "mixin"`,
-                                        pkgVersion = `mkVersion [0]`},
+                                       {pkgName = PackageName "mixin", pkgVersion = mkVersion [0]},
                            pkgUrl = "",
                            setupBuildInfo = Nothing,
                            sourceRepos = [],
