+++ Cabal/tests/ParserTests/regressions/spdx-3.expr
@@ -84,8 +84,7 @@ GenericPackageDescription
                                           (License (ELicense (ELicenseId AGPL_1_0_only) Nothing)),
                            maintainer = "",
                            package = PackageIdentifier
-                                       {pkgName = `PackageName "spdx"`,
-                                        pkgVersion = `mkVersion [0]`},
+                                       {pkgName = PackageName "spdx", pkgVersion = mkVersion [0]},
                            pkgUrl = "",
                            setupBuildInfo = Nothing,
                            sourceRepos = [],
