+++ cabal-install/cabal-install.cabal.pp
@@ -502,7 +502,8 @@ Test-Suite unit-tests
         cabal-lib-client,
         cabal-install-solver-dsl,
         Cabal,
-        Cabal-quickcheck,
+        Cabal-tree-diff,
+        Cabal-QuickCheck,
         containers,
         deepseq,
         mtl,
